+++ src/Kiota.Builder/Refiners/CommonLanguageRefiner.cs
@@ -332,6 +332,22 @@ public abstract class CommonLanguageRefiner : ILanguageRefiner
             KiotaBuilder.AddSerializationMembers(newClass, true, usesBackingStore);
             newClass.Kind = CodeClassKind.Model;
         }
+        // Add the discrimnator function to the wrapper as it will be referenced. 
+        var factoryMethod = newClass.AddMethod(new CodeMethod
+        {
+            Name = "CreateFromDiscriminatorValue",
+            ReturnType = new CodeType { TypeDefinition = newClass, Name = newClass.Name, IsNullable = false },
+            Kind = CodeMethodKind.Factory,
+            IsStatic = true,
+            IsAsync = false,
+        }).First();
+        factoryMethod.AddParameter(new CodeParameter
+        {
+            Name = "parseNode",
+            Kind = CodeParameterKind.ParseNode,
+            Optional = false,
+            Type = new CodeType { Name = "IParseNode", IsExternal = true },
+        });
         return new CodeType {
             Name = newClass.Name,
             TypeDefinition = newClass,